+++ src/components/layout/sign-page/header/SignPageHeader.style.tsx
